refactor(simple-color-picker): move from experimental to components
BREAKING CHANGE: SimpleColorPicker has moved to src/components/simple-color-picker
