+++ src/components/advanced-color-picker/advanced-color-picker.style.js
@@ -1,15 +1,15 @@
 import styled, { css } from "styled-components";
 import { margin } from "styled-system";
 import StyledAdvancedColorPickerCell from "./advanced-color-picker-cell.style";
-import { StyledColorOptions } from "../../__experimental__/components/simple-color-picker/simple-color-picker.style";
-import StyledSimpleColor from "../../__experimental__/components/simple-color-picker/simple-color/simple-color.style";
+import { StyledColorOptions } from "../simple-color-picker/simple-color-picker.style";
+import StyledSimpleColor from "../simple-color-picker/simple-color/simple-color.style";
 import {
   DialogContentStyle,
   DialogInnerContentStyle,
 } from "../dialog/dialog.style";
 import Dialog from "../dialog/dialog.component";
 import StyledIconButton from "../icon-button/icon-button.style";
-import checkerBoardSvg from "../../__experimental__/components/simple-color-picker/color-sample-box/checker-board.svg";
+import checkerBoardSvg from "../simple-color-picker/color-sample-box/checker-board.svg";
 import baseTheme from "../../style/themes/base";
 
 const StyledAdvancedColorPickerWrapper = styled.div`
