+++ src/components/simple-color-picker/color-sample-box/color-sample-box.spec.js
@@ -6,7 +6,7 @@ import StyledColorSampleBox from "./color-sample-box.style";
 import {
   assertStyleMatch,
   expectError,
-} from "../../../../__spec_helper__/test-utils";
+} from "../../../__spec_helper__/test-utils";
 import StyledTickIcon from "../tick-icon/tick-icon.style";
 
 function render(props) {
