+++ src/components/simple-color-picker/color-sample-box/color-sample-box.style.js
