+++ src/components/simple-color-picker/color-sample-box/index.js
