+++ src/components/simple-color-picker/index.d.ts
