+++ src/components/simple-color-picker/index.js
