+++ src/components/simple-color-picker/simple-color-picker.component.js
@@ -7,15 +7,15 @@ import React, {
 } from "react";
 import PropTypes from "prop-types";
 import styledSystemPropTypes from "@styled-system/prop-types";
-import Events from "../../../utils/helpers/events";
-import tagComponent from "../../../utils/helpers/tags/tags";
-import Fieldset from "../../../__internal__/fieldset";
+import Events from "../../utils/helpers/events/events";
+import tagComponent from "../../utils/helpers/tags/tags";
+import Fieldset from "../../__internal__/fieldset";
 import SimpleColor from "./simple-color";
-import RadioButtonMapper from "../../../components/radio-button/radio-button-mapper.component";
+import RadioButtonMapper from "../radio-button/radio-button-mapper.component";
 import { StyledContent, StyledColorOptions } from "./simple-color-picker.style";
-import ValidationIcon from "../../../__internal__/validations/validation-icon.component";
-import { InputGroupContext } from "../../../__internal__/input-behaviour";
-import { filterStyledSystemMarginProps } from "../../../style/utils";
+import ValidationIcon from "../../__internal__/validations/validation-icon.component";
+import { InputGroupContext } from "../../__internal__/input-behaviour";
+import { filterStyledSystemMarginProps } from "../../style/utils";
 
 const marginPropTypes = filterStyledSystemMarginProps(
   styledSystemPropTypes.space
