+++ src/components/simple-color-picker/simple-color-picker.stories.mdx
@@ -1,11 +1,11 @@
 import { useState } from "react";
 import { Meta, Story, Preview, Props } from "@storybook/addon-docs/blocks";
-import StyledSystemProps from "../../../../.storybook/utils/styled-system-props";
+import StyledSystemProps from "../../../.storybook/utils/styled-system-props";
 import { SimpleColorPicker, SimpleColor } from ".";
-import Box from "../../../components/box";
+import Box from "../box";
 
 <Meta
-  title="Experimental/Simple Color Picker"
+  title="Design System/Simple Color Picker"
   parameters={{ info: { disable: true } }}
 />
 
@@ -21,7 +21,7 @@ import Box from "../../../components/box";
 ## Quick Start
 
 ```javascript
-import SimpleColorPicker from "carbon-react/lib/__experimental__/components/simple-color-picker";
+import SimpleColorPicker from "carbon-react/lib/components/simple-color-picker";
 ```
 
 ## Examples
