+++ src/components/simple-color-picker/simple-color-picker.style.js
@@ -1,7 +1,7 @@
 import styled, { css } from "styled-components";
 
-import StyledValidationicon from "../../../__internal__/validations/validation-icon.style";
-import baseTheme from "../../../style/themes/base";
+import StyledValidationicon from "../../__internal__/validations/validation-icon.style";
+import baseTheme from "../../style/themes/base";
 
 const BORDER_WIDTH = 2;
 const getRoundedMaxWidth = (maxWidth, childWidth) =>
