+++ src/components/simple-color-picker/simple-color/index.js
