+++ src/components/simple-color-picker/simple-color/simple-color.d.ts
