+++ src/components/simple-color-picker/simple-color/simple-color.style.js
